+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 * Show Operator name on loan action history. Fixes UIU-163.
 * Rename action names. Fixes UIU-161 and UIU-161.
 * The clear-search button goes back to Users starting state. Fixes UIU-155.
+* Explicitly reset component state in onClearSearch. Fixes UIU-167.
+* Change search placeholder text, and add searching for internal ID and External system ID. Fixes UIU-5.
 * Display Loan Due Date. Upgrade `circulation` dependency to 2.1 and `loan-storage` to 3.1. Fixes UIS-52
 
 ## [2.8.0](https://github.com/folio-org/ui-users/tree/v2.8.0) (2017-07-17)